requirements.txt
These are the necessary libraries to be installed to run the application.

Flask==3.0.3
monai==1.3.1
numpy==1.24.4
pandas==2.0.3
torch==2.3.1+cu121 pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
torchaudio==2.3.1+cu121 
torchvision==0.18.1+cu121
cuda== 11.8 https://developer.nvidia.com/cuda-11-8-0-download-archive
seaborn==0.13.2
pillow==10.3.0
Werkzeug==3.0.3
Python==3.8.10 https://www.python.org/downloads/release/python-3810/
scikit-learn==1.3.2
matplotlib==3.7.5
scipy==1.10.1
ipykernel==6.29.4
